pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "two_tier_document_parser"
version = "0.1.0"
description = "A production-ready, two-tier document parsing API for RAG pipelines"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Daddal" }
]
maintainers = [
    { name = "Daddal" }
]
keywords = [
    "pdf-parser",
    "document-parsing",
    "rag",
    "mineru",
    "pymupdf",
    "fastapi",
    "document-extraction",
    "markdown",
    "vlm",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Markup",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Framework :: FastAPI",
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "loguru>=0.7.2",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "rich>=13.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
fast = [
    "pymupdf4llm",
    "pymupdf",
]
accurate = [
    "torch>=2.0.0",
    "transformers>=4.44.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pypdf>=3.0.0",  # For creating single-page PDFs in tests
]
# Optional telemetry for distributed tracing (API-only pattern)
# API is no-op without SDK - zero overhead for users who don't need tracing
# See: https://opentelemetry.io/docs/concepts/instrumentation/libraries/
telemetry = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
]

[project.urls]
Homepage = "https://github.com/daddal001/two_tier_document_parser"
Documentation = "https://github.com/daddal001/two_tier_document_parser#readme"
Repository = "https://github.com/daddal001/two_tier_document_parser"
Issues = "https://github.com/daddal001/two_tier_document_parser/issues"

[project.scripts]
two-tier-parser-demo = "examples.demo_client:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
known_first_party = ["two_tier_parser"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "mineru.*",
    "pymupdf4llm.*",
    "torch.*",
    "transformers.*",
]
ignore_missing_imports = true
